reorganized file structure into server/client folders
