+++ .gitignore
@@ -1,0 +1,4 @@
+node_modules/
+build/
+config.js
+.DS_Store
