--- Build.txt
@@ -1,37 +1,0 @@
-General Build Notes:
-
-    -- Update ServerConstant.js at deployment to be local or deployment
-
-Building FrontEnd and Backend for Local Builds:
-    -- local.sh
-
-Building FrontEnd on Flip Server:
-
-    -- Set Integrated Terminal to correct location (cs340_project)
-    -- (only needed first time)  npm install
-    -- npm run build
-    -- (only needed first time) npm install -g serve
-    -- serve -s build
-    -- http://flip1.engr.oregonstate.edu:INSERTPORTNUM
-    -- example:
-        http://flip1.engr.oregonstate.edu:36109
-
-Building FrontEnd For Local Builds: 
-
-    -- Builds to localhost:portNum
-    -- To begin the development, run `npm start` or `yarn start`.
-    -- To create a production bundle, use `npm run build` or `yarn build`.
-
-Database Login:
-
-    -- See local notes
-
-Other Misc. Info:
-
-Notice the use of %PUBLIC_URL% in the tags above.
-It will be replaced with the URL of the `public` folder during the build.
-Only files inside the `public` folder can be referenced from the HTML.
-
-Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-work correctly both with client-side routing and a non-root public URL.
-Learn how to configure a non-root public URL by running `npm run build`.
