+++ zillowgame/ServerConfig.mjs
@@ -1,0 +1,6 @@
+export const AddressDeploy = '';
+export const PORT = 5005;
+export const AddressLocal = `http://localhost:${PORT}`;
+export const AddressInUse = AddressLocal;
+
+
