+++ zillowgame/client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "zillowgame",
+  "name": "zillowgame-client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
