+++ zillowgame/client/src/frontend/ValidatePlaylist.mjs
