+++ zillowgame/client/src/frontend/ValidateProperty.mjs
