--- zillowgame/deploy.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-#npm run build
-pm2 serve build 9393
-pm2 start src/backend/DatabaseController.mjs
