--- zillowgame/local.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-node src/backend/DatabaseController.mjs &
-npm start
