+++ zillowgame/server/package.json
@@ -1,0 +1,54 @@
+{
+  "name": "zillowgame-server",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
+    "body-parser": "^1.19.2",
+    "cors": "^2.8.5",
+    "create-react-app": "^5.0.0",
+    "express": "^4.18.2",
+    "forever": "^4.0.3",
+    "madge": "^5.0.1",
+    "mysql": "^2.18.1",
+    "pm2": "^5.2.0",
+    "react": "^17.0.2",
+    "react-burger-menu": "^3.0.6",
+    "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "semantic-ui-react": "^2.1.2",
+    "serve": "^13.0.2",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "node server.js"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "description": "zillowgame-server",
+  "main": "DatabaseController.js",
+  "author": "",
+  "license": "ISC",
+  "type": "module"
+}
