--- zillowgame/src/backend/ServerConstant.js
@@ -1,6 +1,0 @@
-export const AddressDeploy = 'http://flip2.engr.oregonstate.edu:6363';
-export const AddressLocal = 'http://localhost:6363';
-export const SpotifyAddress = 'http://flip3.engr.oregonstate.edu:23401';
-export const AddressInUse = AddressLocal;
-
-
